pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "viki-code"
version = "4.1.4"
description = "Production-oriented swarm coding system with messaging integrations and governed execution"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Rebootix Artificial Intelligence Research and Development" }]
license = "MIT"
license-files = ["LICENSE"]
keywords = ["ai", "coding-agent", "automation", "multi-agent", "developer-tools", "repo-intelligence"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
  "typer>=0.9.0",
  "rich>=13.7.0",
  "textual>=0.52.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.1.0",
  "litellm>=1.40.0",
  "aiofiles>=23.2.0",
  "aiosqlite>=0.19.0",
  "docker>=7.0.0",
  "prometheus-client>=0.19.0",
  "structlog>=23.2.0",
  "tenacity>=8.2.0",
  "platformdirs>=4.0.0",
  "pyyaml>=6.0.1",
  "chardet>=5.2.0",
  "gitignore-parser>=0.1.9",
  "python-json-logger>=2.0.7",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
semantic = ["faiss-cpu>=1.7.4", "sentence-transformers>=2.2.2"]
security = ["bandit>=1.7.5", "semgrep>=1.50.0"]
dev = ["pytest>=8.0.0"]

[project.scripts]
viki = "viki.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["viki*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
